+++ services/galley/src/Galley/Data.hs
@@ -64,6 +64,7 @@ module Galley.Data
     conversationIdsOf,
     conversationMeta,
     conversations,
+    conversationsRemote,
     createConnectConversation,
     createConversation,
     createOne2OneConversation,
@@ -569,6 +570,10 @@ conversationIdsOf ::
   [ConvId] ->
   m [ConvId]
 conversationIdsOf usr cids = runIdentity <$$> retry x1 (query Cql.selectUserConvsIn (params Quorum (usr, cids)))
+
+conversationsRemote :: (MonadClient m) => UserId -> m [Remote ConvId]
+conversationsRemote usr = do
+  (\(d, c) -> toRemote $ Qualified c d) <$$> retry x1 (query Cql.selectUserRemoteConvs (params Quorum (Identity usr)))
 
 createConversation ::
   MonadClient m =>