+++ CHANGELOG.md
@@ -16,7 +16,11 @@
 
 # Unreleased
 
-# Internal changes
+## API Changes
+
+* New endpoint `POST /list-conversations` similar to `GET /conversations`, but which will also return your own remote conversations (if federation is enabled). (#1591)
+
+## Internal changes
 
 * schema-profunctor: add `optField` combinator and corresponding documentation (#1621, #1624).
 * [Federation] Let a receiving backend decide conversation attribute specifics of its users